INCOMPLETE replace pac4j-jee with pac4j-core and make this library an explicit module
To make this library an explicit module (by adding a module-info.java) we face the issue that we depended on pac4j-jee which has several split packages with pac4j-core. That prevents this library from building as an automatic module (by adding Automatic-Module-Name).
By removing pac4j-jee, we circumvent the entire split packages problem here, but we must implement the provided pac4j-core interfaces ourselves.
In doing so, we will need to depend on the javax-servlet-api (see JavalinWebContext).
Since we already depend on javalin, depending on jetty:9, depending on javax-servlet-api, we already have the dependency on the classpath. But javax-servlet-api is an unnamed module (it neither has a module-info.java nor an Automatic-Module-Name). That means, for us to be able to compile, we must make javax-servlet-api an filename-based-module (which maven does for us).
So now we can say, this library "requires javax.servlet.api;". But requiring a filename-based-module is strongly discouraged, resulting in a warning from maven, that we should not publish this artefact to a public repository and ultimately not use any artefact that requires a filename-based-module. (in fact, by trying to include the jar as a dependency on a gradle project, one cannot build the project since there javax.servlet.api will not be found as a module).

So we are stuck with two bad options: (A) require javax.servlet.api (posing the restriction to library users to themselves make javax-servlet-api a filename-based-module) or (B) not depend it an run into the build error "org.pacj4.javalin reads javax.servlet.api.http from unnamed module which it does not require".

One could resolve that by upgrading the jetty-dependency of javalin to now use jakarta-servlet-api (which is an explicit module).